Added destroy_tex_info function
